Add support for skipping the test/suite teardown steps.
Problem:
- We want to be able to run a test and leave its environment intact
  (ie. skip its test/suite teardown steps).
- This option comes in handy when debugging test failures.

Solution:
- Add logic to the test/suite teardown functions that causes them to
  simply return (ie. do nothing) if the user passes the "skip_teardown"
  variable at runtime (and sets its value to "true").
- Sample usage: "py.test --vars skip_teardown:true -- tests/foo.py".
- Note that the pytest-meta plugin adds support for passing variable
  assignments at runtime.

+++ systest/systest_f5mlb/testsuites/marathon/conftest.py
@@ -43,6 +43,8 @@ def default_test_fx(request, marathon, bigip):
     bigip.partition.create(partition, subPath="/")
 
     def teardown():
+        if request.config._meta.vars.get('skip_teardown', None):
+            return
         marathon.apps.delete()
         marathon.deployments.delete()
         bigip.virtual_servers.delete(partition=partition)
@@ -61,6 +63,8 @@ def f5mlb(request, marathon):
     f5mlb = utils.create_f5mlb(marathon)
 
     def teardown():
+        if request.config._meta.vars.get('skip_teardown', None):
+            return
         f5mlb.delete()
 
     request.addfinalizer(teardown)
